% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{add_comments_input}
\alias{add_comments_input}
\title{Input: message content template for add_comments()}
\format{
\subsection{\code{add_comments_input}}{

A data frame with 2 rows and 2 columns representing the message object in API
calls.
}
}
\source{
Contributed by jldc
}
\usage{
add_comments_input
}
\description{
This message content serves as user input for the add_comments function.
}
\keyword{datasets}
