% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anscombe}
\alias{anscombe}
\title{Anscombe's Quartet of ‘Identical’ Simple Linear Regressions}
\format{
A \code{\link[tibble]{tibble}} with 11 observations of 8 variables:
\describe{
  \item{x1}{Vector of integers between 4 and 14.}
  \item{x2}{Same vector as \code{x1}.}
  \item{x3}{Same vector as \code{x1}.}
  \item{x4}{Vector containing 8's and 19's.}
  \item{y1}{Numbers between 3 and 12.5 with mean 7.5 and standard deviation
  2.03 to be analysed together with x1.}
  \item{y2}{Numbers between 3 and 12.5 with mean 7.5 and standard deviation
  2.03 to be analysed together with x2.}
  \item{y3}{Numbers between 3 and 12.5 with mean 7.5 and standard deviation
  2.03 to be analysed together with x3.}
  \item{y4}{Numbers between 3 and 12.5 with mean 7.5 and standard deviation
  2.03 to be analysed together with x4.}
}
}
\source{
Tufte, Edward R. (1989). The Visual Display of Quantitative
  Information, 13–14. Graphics Press.
}
\usage{
anscombe
}
\description{
Four \emph{x-y} data sets which have the same traditional statistical
properties (mean, variance, correlation, regression line, etc.), yet are
quite different.
}
\references{
Anscombe, Francis J. (1973). Graphs in statistical analysis. The
  American Statistician, 27, 17–21. doi: \url{10.2307/2682899}.
}
\keyword{datasets}
