% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain}
\alias{brain}
\title{Brain Data Set}
\format{
A \code{\link[tibble]{tibble}} with 62 rows and 2 variables:
\describe{
  \item{body_weight}{body weight, in kg.}
  \item{brain_weight}{brain weight, in kg.}
}
}
\source{
???
}
\usage{
brain
}
\description{
A data set containing the body weight and brain weight of 62 species.
}
\keyword{datasets}
