% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibRate.R
\name{calibRate}
\alias{calibRate}
\title{Shiny gadget to select exam questions and generate both questions and solutions.}
\usage{
calibRate()
}
\value{
Save the different versions of exams and solutions.
}
\description{
Shiny gadget to select exam questions and generate both questions and solutions.
}
