% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars}
\alias{cars}
\title{Speed and Stopping Distances of Cars}
\format{
A \code{\link[tibble]{tibble}} with 50 observations of 2 variables:
\describe{
  \item{speed}{Speed, in mph.}
  \item{dist}{Stopping distance, in ft.}
}
}
\source{
Ezekiel, M. (1930) Methods of Correlation Analysis. Wiley.
}
\usage{
cars
}
\description{
The data give the speed of cars and the distances taken to stop. Note that
the data were recorded in the 1920s.
}
\references{
McNeil, D. R. (1977) Interactive Data Analysis. Wiley.
}
\keyword{datasets}
