% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cement}
\alias{cement}
\title{A Data Set About Cement Composition}
\format{
A \code{\link[tibble]{tibble}} with 13 observations of 5 variables:
\describe{
  \item{aluminium}{Quantity of aluminium, in grams.}
  \item{silicate}{Quantity of silicate, in grams.}
  \item{aluminium_ferrite}{Quantity of aluminium ferrite, in grams.}
  \item{silicate_bic}{Quantity of silicate bic, in grams.}
  \item{hardness}{Hardness.}
}
}
\source{
???
}
\usage{
cement
}
\description{
A data set containing 13 cements with different compositions leading to
different hardness.
}
\keyword{datasets}
