% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkexam.R
\name{checkexam}
\alias{checkexam}
\title{Shiny gadget to check and analyze results of an exam as exported from gradexam.}
\usage{
checkexam()
}
\value{
A list with all the last run analyses.
}
\description{
Shiny gadget to check and analyze results of an exam as exported from gradexam.
}
\seealso{
genexam()

gradexam()

sendexam()
}
