% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{cloze}
\alias{cloze}
\title{Cloze question}
\usage{
cloze(x)
}
\arguments{
\item{x}{The question's solution.}
}
\description{
Use this function to declare a cloze question in an R Markdown document
with the \code{teachr::teachr_cloze} output format.
}
