% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{codeblock}
\alias{codeblock}
\title{The codeblock for the chapter}
\usage{
codeblock(id, hint)
}
\arguments{
\item{id}{The id of the codeblock}

\item{hint}{The hint.}
}
\description{
The codeblock for the chapter
}
