% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_solutions.R
\name{compare_solutions}
\alias{compare_solutions}
\title{Compare a user's solution to the intended solution}
\usage{
compare_solutions(attempt_file, solution_file, ...)
}
\arguments{
\item{attempt_file}{A character string that represents the path to the file
that contains the user's solution.}

\item{solution_file}{A character string that represents the path to the file
that contains the intended solution.}

\item{...}{Additional arguments to pass to OpenAIR's chat_completion}
}
\value{
A character string that compares the user's solution to the intended
solution and provides feedback to the user.
}
\description{
This function takes a user's solution and compares it to the intended
solution. If the user's solution is incorrect, the function will provide
feedback to the user.
}
\examples{
compare_solutions("R_code/exercise_1.R", "solutions/exercise_1.R")

}
