% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compare_solutions_input}
\alias{compare_solutions_input}
\title{Input: message content template for compare_solutions()}
\format{
\subsection{\code{compare_solutions_input}}{

A data frame with 2 rows and 2 columns representing the message object in API
calls.
}
}
\source{
Contributed by jldc
}
\usage{
compare_solutions_input
}
\description{
This message content serves as user input for the compare_solutions function.
}
\keyword{datasets}
