% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_clean_root.R
\name{course_clean_root}
\alias{course_clean_root}
\title{Remove temporary files in root}
\usage{
course_clean_root()
}
\value{
Clean project root folder
}
\description{
Remove non-essential files and folders from the root of the project. These files are created when .Rmd documents are knited.
}
\author{
Nicolas Mangin
}
