% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_intake_server.R
\name{course_intake_server}
\alias{course_intake_server}
\title{Load intake data}
\usage{
course_intake_server(id, course_data, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
A list containing the intake as a table and a json object.
}
\description{
Module facilitating the loading of an intake (with associated trees, paths, and tests).
}
\author{
Nicolas Mangin
}
