% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_intake_ui.R
\name{course_intake_ui}
\alias{course_intake_ui}
\title{Select and load a intake}
\usage{
course_intake_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A list containing the intake as a table and a json object
}
\description{
Module facilitating the loading of a classification intake.
}
\author{
Nicolas Mangin
}
