% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_load_server.R
\name{course_load_server}
\alias{course_load_server}
\title{Load course data}
\usage{
course_load_server(id, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
A list of course data.
}
\description{
Module facilitating the loading of all relevant course data.
}
\author{
Nicolas Mangin
}
