% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_load_ui.R
\name{course_load_ui}
\alias{course_load_ui}
\title{Load course data}
\usage{
course_load_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A list of course data.
}
\description{
Module facilitating the loading of all relevant course data.
}
\author{
Nicolas Mangin
}
