% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_references_server.R
\name{course_references_server}
\alias{course_references_server}
\title{Load course references}
\usage{
course_references_server(id, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Tibble with references.
}
\description{
Module facilitating the loading of all relevant references for the course.
}
\author{
Nicolas Mangin
}
