% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_references_ui.R
\name{course_references_ui}
\alias{course_references_ui}
\title{Load course references}
\usage{
course_references_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Tibble with references.
}
\description{
Module facilitating the loading of all relevant references for the course.
}
\author{
Nicolas Mangin
}
