% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_test_ui.R
\name{course_test_ui}
\alias{course_test_ui}
\title{Select and load a test}
\usage{
course_test_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A list containing a test and a test folder path used in testR, gradR, and reportR.
}
\description{
Module facilitating the selection of a test to edit, grade, and feedback.
}
\author{
Nicolas Mangin
}
