% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_tree_server.R
\name{course_tree_server}
\alias{course_tree_server}
\title{Load tree data}
\usage{
course_tree_server(id, course_data, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
A list which can be treated as a sTreeR object.
}
\description{
Module facilitating the loading of a js classification tree.
}
\author{
Nicolas Mangin
}
