% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_tree_ui.R
\name{course_tree_ui}
\alias{course_tree_ui}
\title{Select and load a tree}
\usage{
course_tree_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A list containing the tree as a table and a json object
}
\description{
Module facilitating the loading of a classification tree.
}
\author{
Nicolas Mangin
}
