% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_update_server.R
\name{course_update_server}
\alias{course_update_server}
\title{Update documents and data}
\usage{
course_update_server(id, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Nothing. All changes ore on disk
}
\description{
Module facilitating the update of underlying course data.
}
\author{
Nicolas Mangin
}
