% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_update_ui.R
\name{course_update_ui}
\alias{course_update_ui}
\title{Update documents and data}
\usage{
course_update_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A list of folder paths based on a standard course structure and used by the application.
}
\description{
Module facilitating the update of underlying course data.
}
\author{
Nicolas Mangin
}
