% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assignment.R
\name{create_assignment}
\alias{create_assignment}
\title{Create a new assignment}
\usage{
create_assignment(name, number, week)
}
\arguments{
\item{name}{Assignment name as a character string. If left blank, assignments
are titled as 'Assignment (number)'}

\item{number}{Assignment number as an integer value.}

\item{week}{Week/module number for the assignment. The assignment is then
automatically placed in the correct location on the course website.}
}
\value{

}
\description{
Creates a new assignment for the course and automatically updates the course website,
GitHub repositories, and Canvas.
}
