% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assignment_assignments.R
\name{create_assignment_assignments}
\alias{create_assignment_assignments}
\title{Create a new assignment in the 'assignments' repository}
\usage{
create_assignment_assignments(name, number, week)
}
\arguments{
\item{name}{See create_assignment()}

\item{number}{See create_assignment()}

\item{week}{See create_assignment()}
}
\description{
Internal function. Creates a new assignment in the 'assignments' repository.
}
