% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assignment_canvas.R
\name{create_assignment_canvas}
\alias{create_assignment_canvas}
\title{Create a new assignment on Canvas}
\usage{
create_assignment_canvas(name)
}
\arguments{
\item{name}{Assignment name as a character string.}
}
\description{
Internal function. Creates a new assignment for the course on Canvas.
}
