% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assignment_website.R
\name{create_assignment_website}
\alias{create_assignment_website}
\title{Create a new assignment in the website repository}
\usage{
create_assignment_website(name, number, week)
}
\arguments{
\item{name}{See create_assignment()}

\item{number}{See create_assignment()}

\item{week}{See create_assignment()}
}
\description{
Updates assignment information in the website repository for the corresponding
module. Needs to be run from the website repository.
}
