% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lecture.R
\name{create_lecture}
\alias{create_lecture}
\title{Create a new lecture}
\usage{
create_lecture(name, day)
}
\arguments{
\item{name}{Lecture name as a character string. Example: 'Linear Regression: The Basics'}

\item{day}{Day/lecture number as an integer value. For example, for courses
with lectures on M/W/F, each new week would be a new module and day = 1, 2, 3
for M, W, F, respectively.}
}
\description{
Creates a new lecture for the course and automatically updates the course website,
GitHub repositories, and Canvas.
}
