% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lecture_canvas.R
\name{create_lecture_canvas}
\alias{create_lecture_canvas}
\title{Create a new lecture on Canvas}
\usage{
create_lecture_canvas(name, week, day)
}
\arguments{
\item{name}{See create_lecture()}

\item{week}{See create_lecture()}

\item{day}{See create_lecture()}
}
\description{
Internal function. Creates a new lecture on Canvas.
}
\keyword{internal}
