% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lecture_lectures.R
\name{create_lecture_lectures}
\alias{create_lecture_lectures}
\title{Create a new lecture in the 'lectures' repository}
\usage{
create_lecture_lectures(name, week, day)
}
\arguments{
\item{name}{See create_lecture()}

\item{week}{See create_lecture()}

\item{day}{See create_lecture()}
}
\description{
Internal function. Creates a new lecture in the 'lectures' repository.
}
\keyword{internal}
