% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lecture_pdf.R
\name{create_lecture_pdf}
\alias{create_lecture_pdf}
\title{Create PDF out of lecture notes}
\usage{
create_lecture_pdf(week, day)
}
\arguments{
\item{week}{See create_lecture()}

\item{day}{See create_lecture}
}
\description{
Creates a PDF out of the lecture notes HTML file. Google Chrome must be installed
in order to use this funciton.
}
