% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lecture_website.R
\name{create_lecture_website}
\alias{create_lecture_website}
\title{Create a new lecture in the website repository}
\usage{
create_lecture_website(name, day)
}
\arguments{
\item{name}{See create_lecture()}

\item{day}{See create_lecture()}
}
\description{
Internal function. Creates a new lecture in the 'lectures' repository.
}
