% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module.R
\name{create_module}
\alias{create_module}
\title{Create a new module}
\usage{
create_module(name, week)
}
\arguments{
\item{name}{Module name as a character string. Example: 'Linear Regression'}

\item{week}{Module/week number as an integer value. Note that in this package
module number = week number as these are assumed to be the same.}
}
\description{
Creates a new module for the course and automatically updates the course website,
GitHub repositories, and Canvas.
}
