% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module_canvas.R
\name{create_module_canvas}
\alias{create_module_canvas}
\title{Create a new module on Canvas}
\usage{
create_module_canvas(name, week)
}
\arguments{
\item{name}{See create_module()}

\item{week}{See create_module()}
}
\description{
Internal function. Creates a new module on Canvas.
}
\keyword{internal}
