% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module_lectures.R
\name{create_module_lectures}
\alias{create_module_lectures}
\title{Create a new module in the 'lectures' repository}
\usage{
create_module_lectures(week)
}
\arguments{
\item{week}{See create_module()}
}
\description{
Internal function. Creates a new module in the 'lectures' repository.
}
