% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module_website.R
\name{create_module_website}
\alias{create_module_website}
\title{Create a new module on the course website}
\usage{
create_module_website(name, week)
}
\arguments{
\item{name}{See create_module()}

\item{week}{See create_module()}
}
\description{
Creates a new module on the course website. Note that this function needs to
be run from the "(organization).github.io" repository.
}
