% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_variable.R
\name{decrypt_variable}
\alias{decrypt_variable}
\title{Decrypt a variable}
\usage{
decrypt_variable(x, var, key)
}
\arguments{
\item{x}{Tibble. Data in with the variable to be decrypted}

\item{var}{Character. Name of the variable which should be decrypted}

\item{key}{Character. String used to decrypt the variable}
}
\value{
Tibble. Initial tibble in which the specified variable is decrypted
}
\description{
Decrypt the specified variable of a data frame using the specified key
}
\author{
Nicolas Mangin
}
