% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_variable.R
\name{encrypt_variable}
\alias{encrypt_variable}
\title{Encrypt a variable}
\usage{
encrypt_variable(x, var, key)
}
\arguments{
\item{x}{Tibble. Data in with the variable to be encrypted}

\item{var}{Character. Name of the variable which should be encrypted}

\item{key}{Character. String used to encrypt the variable}
}
\value{
Tibble. Initial tibble in which the specified variable is encrypted
}
\description{
Encrypt the specified variable of a data frame using the specified key
}
\author{
Nicolas Mangin
}
