% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_code.R
\name{explain_code}
\alias{explain_code}
\title{Explain Code using OpenAI's chat_completion function}
\usage{
explain_code(code, ...)
}
\arguments{
\item{code}{A character string that represents the code that the user wants
to explain.}

\item{...}{Additional arguments to pass to OpenAIR's chat_completion}
}
\value{
A character string that represents the explanation for the code.
}
\description{
This function takes a code as input and uses OpenAIR's chat_completion
function to generate an explanation for the code.
}
\examples{
explain_code("for(i in 1:10) { print(i) }")

}
