% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_app.R
\name{export_app}
\alias{export_app}
\title{Export teaching apps provided by \code{\link[<teachR>:<teachR>-package]{teachR}}}
\usage{
export_app(app, dest)
}
\arguments{
\item{app}{The name of the app to run.}

\item{dest}{The destination of the app folder on the user's computer. A folder
with the name of the app will be created at this location.}
}
\value{
This function does not return anything.
}
\description{
This function expor a Shiny app provided by the \code{\link[<teachR>:<teachR>-package]{teachR}}
library to a local folder on a user's computer. Useful for modifying the app
or running it on a Shiny Server instance for example.
}
\examples{
# Export app to the current working directory
export_app("aggregation_segregation", ".")
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

