% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_documents_count.R
\name{filter_documents_count}
\alias{filter_documents_count}
\title{Count the different kinds of selected documents.}
\usage{
filter_documents_count(selection)
}
\arguments{
\item{selection}{Tibble. Filtered documents.}
}
\value{
Character. String reporting the number of each document after filter application.
}
\description{
Function reporting the number of documents of each kind after applying filters.
}
\author{
Nicolas Mangin
}
