% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_languages_server.R
\name{filter_languages_server}
\alias{filter_languages_server}
\title{Filter documents based on languages.}
\usage{
filter_languages_server(id, course_data)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}
}
\value{
Tibble. list of documents existing in all the selected languages.
}
\description{
Module allowing the user to find select document based on whether they exist in languages other than the main one.
}
\author{
Nicolas Mangin
}
