% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_make_ui.R
\name{filter_make_ui}
\alias{filter_make_ui}
\title{Create filters UI}
\usage{
filter_make_ui(ns, preselection, filter_variables, tags = NA)
}
\arguments{
\item{ns}{Function. Apply the ID of the module in which the function is embedded.}

\item{preselection}{Tibble}

\item{filter_variables}{Tibble. Output of the function filter_prepare_variables.}

\item{tags}{Tibble. List of tags.}
}
\value{
A user interface with adequate filters for each variable.
}
\description{
Function creating user interfaces to select documents based on specified variables
}
\seealso{
filter_prepare_variables

filter_tibble
}
\author{
Nicolas Mangin
}
