% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_prepare_variables.R
\name{filter_prepare_variables}
\alias{filter_prepare_variables}
\title{Prepare a list of variables for filters}
\usage{
filter_prepare_variables(preselection, filter_family)
}
\arguments{
\item{preselection}{Tibble. List of remaining documents after some prior selection.}

\item{filter_family}{Character. Whether the filtering step is for "common_tags" "custom_tags", "ratings", "views", or "results".}
}
\value{
A tibble indicating the variables for which filters should be created, the associated input id, and the format of the filter. Used as input for filter_make_ui.
}
\description{
Function creating the variable list serving as input to the function making filters.
}
\seealso{
filter_make_ui

filter_tibble
}
\author{
Nicolas Mangin
}
