% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_statistics_server.R
\name{filter_statistics_server}
\alias{filter_statistics_server}
\title{Select documents from statistics (server)}
\usage{
filter_statistics_server(id, course_data)
}
\arguments{
\item{id}{ID of the module to connect the user interface to the appropriate server side.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}
}
\value{
Tibble. List of selected documents.
}
\description{
Module allowing the user to select documents by clicking on the different sections of a classification statistics.
}
\author{
Nicolas Mangin
}
