% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_statistics_ui.R
\name{filter_statistics_ui}
\alias{filter_statistics_ui}
\title{Select documents from statistics}
\usage{
filter_statistics_ui(id)
}
\arguments{
\item{id}{ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Tibble. List of selected documents.
}
\description{
Module allowing the user to select documents based on their statistics.
}
\author{
Nicolas Mangin
}
