% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tags_ui.R
\name{filter_tags_ui}
\alias{filter_tags_ui}
\title{Select documents from tags}
\usage{
filter_tags_ui(id)
}
\arguments{
\item{id}{ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Tibble. List of selected documents.
}
\description{
Module allowing the user to select documents by clicking on their tags.
}
\author{
Nicolas Mangin
}
