% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tibble.R
\name{filter_tibble}
\alias{filter_tibble}
\title{Filter tibble}
\usage{
filter_tibble(dataset, variable, filter_value = NULL, filter_type = "pattern")
}
\arguments{
\item{dataset}{Tibble. Data to be filtered.}

\item{variable}{Character. Variable used for filtering.}

\item{filter_value}{Character. Value to be filtered.}

\item{filter_type}{Character. Type of filter applied. One of the following: "logical" for a true or false, selection" for a single string value, "multiple" for multiple string values, "pattern" for a regular expression, "range" for a range of numeric values, "value", for a single value.}
}
\value{
Tibble after application of the relevant filter.
}
\description{
Function applying different kinds of filters on a tibble. The procedure applied is determined by the filter type.
}
\author{
Nicolas Mangin
}
