% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tree_server.R
\name{filter_tree_server}
\alias{filter_tree_server}
\title{Select documents from tree}
\usage{
filter_tree_server(id, intake, course_data)
}
\arguments{
\item{id}{ID of the module to connect the user interface to the appropriate server side.}

\item{intake}{Reactive. Function containing all information about the selected intake.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}
}
\value{
Tibble. List of selected documents.
}
\description{
Module allowing the user to select documents by clicking on the different sections of a classification tree.
}
\author{
Nicolas Mangin
}
