% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tree_ui.R
\name{filter_tree_ui}
\alias{filter_tree_ui}
\title{Select documents from tree (UI)}
\usage{
filter_tree_ui(id)
}
\arguments{
\item{id}{ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Tibble. List of selected documents.
}
\description{
Module allowing the user to select documents by clicking on the different sections of a classification tree.
}
\author{
Nicolas Mangin
}
