% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_index.R
\name{fun.index}
\alias{fun.index}
\title{Create html index of used functions}
\usage{
fun.index(
  folder,
  output = "fun_index",
  course = "other",
  url = "",
  pattern = ".Rmd",
  recursive = T
)
}
\arguments{
\item{folder}{\code{character}. Path to folder containing lab sheet/practical .Rmd files.}

\item{output}{\code{character}. Name of output HTML file without extention. "fun_index" by default}

\item{course}{\code{character}. Course name to govern colour scheme. One of \code{"dapR_1", "daprR_2", "dapR_3", "usmr", "msmr", "other"}. \code{"other"} by default.}

\item{url}{\code{character}. URL paths (absolute or relative) to find reference HTML practicals/sheets. Either length 1 or length equal to number of files indexed. By default \code{""}.}

\item{pattern}{\code{character}. Pattern to match when searching.}

\item{recursive}{\code{logical}. Should subfolders of \code{folder=} be searched? \code{TRUE} by default.}
}
\description{
Function creates an index files of all functions inside all code chunks within .Rmd files in a given folder (optionally only those conforming to a pattern, see Arguments). The index file is a HTML file with links to individual lab sheets/practicals.
}
\details{
It's best to keep website structure flat. Put all practicals/lab sheets in a single folder and leave \code{url=} at default. fun_index.html should be placed in the same folder. If this is not the case, \code{url=} must be change otherwise links will be broken.
}
\examples{
fun.index("N:/teaching/analysing_data", , pattern = "practical\\\\d+\\\\.Rmd")
}
