% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_questions.R
\name{generate_questions}
\alias{generate_questions}
\title{Generate Questions Function}
\usage{
generate_questions(
  sample_questions,
  n_questions,
  additional_instructions = "",
  ...
)
}
\arguments{
\item{sample_questions}{Character string or path to text file. Sample questions or path to a text file containing the sample questions.}

\item{n_questions}{Numeric. The number of questions to generate.}

\item{additional_instructions}{Character string. Additional instructions to guide the question generation. Default is an empty string.}
}
\value{
A data frame containing the generated questions.
}
\description{
This function generates a specified number of questions based on provided sample questions.
The questions are generated using OpenAI's GPT model. This function can also accept a text file as input for the sample questions.
}
\details{
The function begins by checking if the \code{sample_questions} parameter is a path to an existing file. If so, it reads the lines from the file, otherwise, it assumes that \code{sample_questions} is a string of sample questions.
The function then processes the user input to create an appropriate input for the OpenAI chat model. After the chat process, it calculates and displays the total number of tokens used in the generation process.
Lastly, it processes the response from the chat model to extract the generated questions, which it then returns.
}
\examples{
\dontrun{
generate_questions("What is R?", 5, "Make the questions intermediate level.")
}

}
\seealso{
\url{https://openai.com/research/openai-chatgpt/}
}
