% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_structure.R
\name{generate_structure}
\alias{generate_structure}
\title{Generate the structure of a question bank}
\usage{
generate_structure()
}
\description{
Recompose structural databases of a question bank.
}
\author{
Nicolas Mangin
}
