% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genexam.R
\name{genexam}
\alias{genexam}
\title{Shiny gadget to select exam questions and generate both questions and solutions.}
\usage{
genexam()
}
\value{
Save the different versions of exams and solutions.
}
\description{
Shiny gadget to select exam questions and generate both questions and solutions.
}
\seealso{
gradexam()

checkexam()

sendexam()
}
