% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTest.R
\name{genTest}
\alias{genTest}
\title{Generate Tests}
\usage{
genTest()
}
\value{
Create an exam folder will all the the necessary files fo test administration and subsequent steps.
}
\description{
Shiny gadget to select test questions and generate both tests and solutions for several delivery formats.
}
\author{
Nicolas Mangin
}
