% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quick_com.R
\name{get.quick.com}
\alias{get.quick.com}
\title{Mark coursework assignment Rmd}
\usage{
get.quick.com(csv_path, rubric, output_dir, warn_col = "#cc0000", code_col = "#b38ed2", tabbed = T, keep_rmd = F)
}
\arguments{
\item{csv_path}{\code{character}. Path or URL of a csv file from which to extract quick comments.}

\item{rubric}{\code{character}. Sourceable file that produces a rubric list object from which to get colours of buttons so that they match colours of rubric criteria from \code{mark()}. See URL in example.}

\item{output_dir}{\code{character}. Which study to get correct results for? Currently either \code{"red"} or \code{"green"}.}

\item{warn_col}{\code{numeric}. Hex code of colour for general comments.}

\item{code_col}{\code{numeric}. Hex code of colour for comments in code.}

\item{tabbed}{\code{logical}. Should sections be tabs (\code{TRUE}; default) or FAQ-style buttons (HTML \code{<details>})?}

\item{keep_rmd}{\code{logical}. Should the source .Rmd file be kept (in working directory)? \code{FALSE} by default.}
}
\value{
Function produces a HTML document saved in \code{output_dir}.
}
\description{
For Rmd submissions of reports and other coursework. Function checks word count and inserts space for feedback.
}
\examples{
get.quick.com(csv_path = "https://docs.google.com/spreadsheets/d/19CXCZk28CQzX4MzQ86a5U-ijYQvP47x0mfX7pQk6KKo/export?gid=588770828&format=csv",
rubric = "https://raw.githubusercontent.com/SussexPsychMethods/and_pub/master/marking/sussex_rubric.R",
output_dir = "../mivalek_io/adata/marking/")
}
