% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submitted.R
\name{get.submitted}
\alias{get.submitted}
\title{Wrapper for \code{rcanvas::get_submissions}}
\usage{
get.submitted(
  token,
  domain = "https://canvas.sussex.ac.uk",
  course_id,
  assign_id
)
}
\arguments{
\item{token}{\code{character}. Canvas access token.}

\item{domain}{\code{character}. Canvas domain.}

\item{course_id}{\code{numeric}. ID of Canvas course/module from which to retrieve assignment}

\item{assign_id}{\code{numeric}. ID code of Canvas assignment to retrieve.}
}
\description{
Retrieves assignment submissions from Canvas from only those students who  have submitted .
}
\examples{
submitted <- get.submitted("1204923CwEH1wk2GKcJqbXF3Nja3QUg7G2Ksr6dBscIN9", course_id = 1234, assign_id = 1324)
}
