% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_output}
\alias{get_output}
\title{Write console output to clipboard}
\usage{
get_output(x)
}
\arguments{
\item{x}{Valid R command whose output you want to copy.}
}
\description{
Captures console output of a command (\code{show()}) and copies it to clipboard as a HTML code block.
}
\examples{
x <- rnorm(100)
y <- 4 + .7 * x + rnrom(100)
get_output(summary(lm(y ~ x)))
}
