% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_data.R
\name{get_pkg_data}
\alias{get_pkg_data}
\title{Retrieve data from packages}
\usage{
get_pkg_data(pkgname, dataset)
}
\arguments{
\item{pkgname}{Character. Name of the package containing the data.}

\item{dataset}{Character. Name of the data to be retrieved from this package}
}
\value{
data
}
\description{
Get the requested data from a specified package.
}
\author{
Nicolas Mangin
}
