% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_path.R
\name{get_site_path}
\alias{get_site_path}
\title{Retrieve the path to the "inst/site" folder}
\usage{
get_site_path()
}
\description{
Internal function. Retrieves the path to the "inst/site" folder.
}
\keyword{internal}
