% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradEssay.R
\name{gradEssay}
\alias{gradEssay}
\title{Shiny gadget to facilitate grading open questions and essays.}
\usage{
gradEssay()
}
\value{
Three .csv files: one with students' points, one with updated solutions, one with updated criterias.
}
\description{
Shiny gadget to facilitate grading open questions and essays.
}
