% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradexam.R
\name{gradexam}
\alias{gradexam}
\title{Shiny gadget to gather informations about questions, participants, and answers.
Then you can quickly analyze results across versions and questions.}
\usage{
gradexam()
}
\description{
Shiny gadget to gather informations about questions, participants, and answers.
Then you can quickly analyze results across versions and questions.
}
\seealso{
genexam()

checkexam()

sendexam()
}
