% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_canvas.R
\name{initialize_canvas}
\alias{initialize_canvas}
\title{Initialize Canvas settings for 'teachr' course}
\usage{
initialize_canvas(name, organization, path)
}
\arguments{
\item{name}{See create_course()}

\item{organization}{See create_course()}

\item{path}{See create_course()}
}
\description{
Internal function. Creates a syllabus and home page for Canvas. Only run this
function once.
}
