% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_directories.R
\name{initialize_directories}
\alias{initialize_directories}
\title{Initialize directories for 'teachr' course}
\usage{
initialize_directories(name, organization, path)
}
\arguments{
\item{name}{See create_course()}

\item{organization}{See create_course()}

\item{path}{See create_course()}
}
\description{
Internal function. Creates the appropriate directories at 'path'.
}
\keyword{internal}
