% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_website.R
\name{initialize_website}
\alias{initialize_website}
\title{Initialize website for 'teachr' course}
\usage{
initialize_website(name, organization, path)
}
\arguments{
\item{name}{See create_course()}

\item{organization}{See create_course()}

\item{path}{See create_course()}
}
\value{

}
\description{
Internal function. Creates the website. Only run this function once.
}
