% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_funs.R
\name{insertSol}
\alias{insertSol}
\title{RStudio addins for inserting teaching material divs and quiz chunks}
\usage{
insertSol()
}
\description{
RStudio addins for inserting teaching material divs and quiz chunks
}
