% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{job}
\alias{job}
\title{Job Data Set}
\format{
A \code{\link[tibble]{tibble}} with 20 observations of 3 variables:
\describe{
  \item{average_score}{Average score.}
  \item{years_service}{Years of service.}
  \item{sex}{Gender.}
}
}
\source{
???
}
\usage{
job
}
\description{
A data set about jobs.
}
\keyword{datasets}
