% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_button.R
\name{link.button}
\alias{link.button}
\title{Link a button on course content page to URL}
\usage{
link.button(course, sem = NULL, weeks = "all", label, links)
}
\arguments{
\item{course}{character. Course page to modify: one of "dapR_1", "daprR_2", "dapR_3", "usmr", or "msmr".}

\item{sem}{numeric. Semester to modify: either 1 or 2. Only applicable for 2-semester courses, currently dapR 1 and 2.}

\item{weeks}{Weeks in which to link buttons. Defaults to "all", otherwise numeric vector.}

\item{label}{character. Name of the button to link.}

\item{links}{character vector. URLs for the buttons. Either length 1 or length(weeks)}
}
\value{
Function does not return anything of use.
}
\description{
Function links a given button (all weeks by default) to URLs specified in links=.
}
\details{
IMPORTANT: Set working directory to website folder!.
}
\examples{
# Link Explainer button for each week to 10 different URLs
link.button(course = "dapR_1", sem = 2, label = "Explainer",
link = paste0("/explainers/topic_", sprintf("\%02d", 1:10), ".html"))
}
