% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_canvas.R
\name{link_canvas}
\alias{link_canvas}
\title{Link to a Canvas course}
\usage{
link_canvas(token, course_id)
}
\arguments{
\item{token}{Canvas API token number. Can be obtained in Settings > Approved Integrations > New Access Token.}

\item{course_id}{Canvas course number.}
}
\value{

}
\description{
Establish a connection with the Canvas Instructure LMS API. Doing so allows for automatic updates to the Canvas course. This only needs to be performed once.
}
