% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_categories.R
\name{list_categories}
\alias{list_categories}
\title{List the categories of all apps available through \code{\link[<teachR>:<teachR>-package]{teachR}}}
\usage{
list_categories()
}
\value{
A vector of character strings, each of them corresponding to a different
category.
}
\description{
This function generates a listing  of all the categories of apps available through
the \code{\link[<teachR>:<teachR>-package]{teachR}} library.
}
\examples{
list_categories()
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

