% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_keywords.R
\name{list_keywords}
\alias{list_keywords}
\title{List the keywords of all apps available through \code{\link[<teachR>:<teachR>-package]{teachR}}}
\usage{
list_keywords()
}
\value{
A vector of character strings, each of them corresponding to a different
keyword.
}
\description{
This function generates a listing of all the keywords of apps available through 
the \code{\link[<teachR>:<teachR>-package]{teachR}} library.
}
\examples{
list_keywords()

}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

