% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_paper.R
\name{make.paper}
\alias{make.paper}
\title{Create formatted exam paper in Word, PDF, or HTML}
\usage{
make.paper(
  items,
  code_file = NULL,
  output_format = "word",
  out_file_name,
  yaml_header,
  title_page,
  ref_file = NULL,
  pre_preocessor = NULL,
  randomise_qs = TRUE,
  randomise_response_opts = TRUE,
  sample = FALSE
)
}
\arguments{
\item{code_file}{\code{character}. Path to .Rmd file(s) with code chunks.}

\item{output_format}{\code{character}. Output format for paper: one of \code{"word", "pdf", "html"}.}

\item{out_file_name}{\code{character}. Name of the output file without extension.}

\item{yaml_header, }{@param title_page \code{character}. Vectors including lines for title page and YAML front matter. It is recommended to generate these using a function similar to \code{\link{sussex.boilerplate}()}.}

\item{ref_file}{\code{character}. Path to a reference docx file with formatting and headers. One can be found in the package folder.}

\item{pre_preocessor}{Pre-processor function to include \code{title_page} in the final document. Currently uses \code{teachR:::sus_pre_processor()} that gets called by \code{teachR:::exam_paper_document()} passed to \code{rmarkdown::render()}. Using alternatives not yet implemented.}

\item{randomise_qs}{\code{logical}. Should order of items be randomised (multi-items that share a single stem/paradigm will be appended to the end of the exam in either case). \code{TRUE} by default.}

\item{randomise_response_opts}{\code{TRUE} to generate a sample paper with correct responses at the end of the paper. \code{FALSE} by default.}

\item{items.}{Data frame of items for a given exam paper. It is recommended to first generate a question bank with \code{\link{question.bank}()} and then select individual paper items.}
}
\value{
Function returns a message if successful and, if \code{sample=FALSE}, produces two versions of paper; one with correct answers (..._WITH_KEY in file name) and one without. A scoring key CSV file gets also generated. If \code{sample=TRUE},, only the sample paper gets generated.
}
\description{
Requires a tibble of exam items, such as the one returned by \code{\link{question.bank}()} and can take one or more .Rmd files with code chunks with code required to generate item data/tables/figures etc.
}
\details{
Function requires a .css and .js files for correct formatting of lab sheets/handouts. These files sit on the stats website in the [root]/sheet_files folder and the path is hard-coded into the function. Look for css and js objects in function body.
}
\examples{
make.paper(paper_items, "exam_code.Rmd", out_file_name = "AnD_MCQ_paper_19_20",
           yaml_header = boilerplate$yaml, title_page = boiler_plate$title_page,
           ref_file = "paper_ref_doc.docx", pre_processor = teachR:::sus_pre_processor)
}
