% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scan_A4.R
\name{make_scan_A4}
\alias{make_scan_A4}
\title{Simplification of the function exams2nops designed by Zeileis et al. to produce only the answer sheet and database.}
\usage{
make_scan_A4(exam, name = "scan", language = "en",
  title = "Title of the exam", institution = "Academic Institution",
  date = "2018-05-10", startid = 1, alternatives = 5,
  reglength = 7L, encoding = "")
}
\arguments{
\item{exam}{Output of the function exams2pdf.}

\item{name}{character. A name prefix for resulting exams and RDS file.}

\item{language}{character. Path to a DCF file with a language specification. Currently, the package ships: English ("en"), Dutch ("nl"), French ("fr"), German ("de"), Hungarian ("hu"), Italian ("it"), Romanian ("ro"), Portuguese ("pt"), Spanish ("es").}

\item{title}{character. Title of the exam, e.g., "Introduction to Statistics".}

\item{institution}{character. Name of the institution at which the exam is conducted.}

\item{date}{character or "Date" object specifying the date of the exam.}

\item{startid}{integer. Starting ID for the exam numbers (defaults to 1).}

\item{alternatives}{integer. Number of boxes.}

\item{reglength}{integer. Number of digits in the registration ID. The default is 7 and it can be increased up to 10.}

\item{encoding}{character. encoding}
}
\value{
Answer sheet for the exam and database of solutions.
}
\description{
Simplification of the function exams2nops designed by Zeileis et al. to produce only the answer sheet and database.
}
\references{
Zeileis, A., N. Umlauf, and F. Leisch. 2014. Flexible Generation of E-Learning Exams in R: Moodle Quizzes, OLAT Assessments, and Beyond. Journal of Statistical Software 58 (1).
}
