% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark}
\alias{mark}
\title{Mark coursework assignment Rmd}
\usage{
mark(file, feedback = F, count_words = !feedback, limit, color = "#b38ed2")
}
\arguments{
\item{file}{\code{character}. Path to .Rmd file to mark.}

\item{feedback}{\code{logical}. Should feedback box with marker's comments appear at the bottom of document? \code{FALSE} by default.}

\item{count_words}{\code{logical}. \code{TRUE} counts body text words and if word count exceeds \code{limit}, inserts a warning line in HTML. \code{!feedback} by default.}

\item{limit}{\code{numeric}. Word count limit}

\item{color}{\code{character}. Single valid colour string (hex code or any of the values in \code{colours()}).}
}
\value{
\code{TRUE} if output .Rmd and .html file were successfully created.
}
\description{
For Rmd submissions of reports and other coursework. Function checks word count and inserts space for feedback.
}
\details{
Function run with \code{feedback = FALSE} will overwrite the original Rmd file and knit it into HTML. This should be done before marking in order to automate word count. The overwritten Rmd should then be used for comments and feedback. Once done, \code{mark(feedback = TRUE)} should be run on the edited Rmd. This will output a ..._marked.html file that can be returned to students.

In-text comments can be inserted into .Rmd file on a lew line surrounded by the <fb></fb> HTML tag.
}
\examples{
# first run
mark("C:/work/201000.Rmd")
# then
mark("C:/work/201000.Rmd", T)
}
