% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter.R
\name{mc_opts}
\alias{mc_opts}
\title{Mulitple choice options for the chapter}
\usage{
mc_opts(..., correct = NULL)
}
\arguments{
\item{...}{A name-value pair where name shows the text, value shows the
message. Each pair should be an option.}

\item{correct}{The number of the option that is correct.}
}
\description{
Mulitple choice options for the chapter
}
