% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{mchoice}
\alias{mchoice}
\title{Multiple choice question}
\usage{
mchoice(opts, sol)
}
\arguments{
\item{opts}{The options to choose from (character vector)}

\item{sol}{The correct solution (the correct value from \code{opts}, or the
position of the correct value)}
}
\description{
Use this function in a cloze inline chunk to describe a single choice
question.
}
\details{
Note: this is only supported for exams 2.4. CRAN version is at 2.3-6.
Install the latest from R-Forge.
\code{install.packages("exams", repos = "http://R-Forge.R-project.org")}
}
