% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mussels}
\alias{mussels}
\title{Mussels Data Set}
\format{
A \code{\link[tibble]{tibble}} with 82 observations of 5 variables:
\describe{
  \item{length}{Mussel length, in mm.}
  \item{width}{Mussel width, in mm.}
  \item{height}{Mussel height, in mm.}
  \item{size}{Mussel size, in mm3.}
  \item{weight}{Mussel weight, in kg.}
}
}
\source{
???
}
\usage{
mussels
}
\description{
A data set about mussels.
}
\keyword{datasets}
