% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_quest.R
\name{param_quest}
\alias{param_quest}
\title{Function to retrieve or generate parameters used when questions are generated.}
\usage{
param_quest(wdir = "", question_id = "")
}
\arguments{
\item{wdir}{Character. Working directory.}

\item{question_id}{Character. ID of the question.}
}
\value{
Parameters for the questions.
}
\description{
Function to retrieve or generate parameters used when questions are generated.
}
