% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_func.R
\name{parse_qus}
\alias{parse_qus}
\title{Functions used for within \code{\link{make.paper}()} and \code{\link{question.bank}()}.}
\usage{
parse_qus(x, group = F, author)
}
\description{
Functions used for within \code{\link{make.paper}()} and \code{\link{question.bank}()}.
}
