% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{Preview marked HTML report in RStudio Viewer}
\usage{
preview(
  file,
  show_rmd = F,
 
    rubric_url = "https://raw.githubusercontent.com/SussexPsychMethods/and_pub/master/marking/sussex_rubric.R",
 
    fdbck_boiler_text = "https://raw.githubusercontent.com/SussexPsychMethods/and_pub/master/marking/fdbck_boilerplate.txt",
  install_missing_pkgs = F,
 
    rubric_btn_url = "https://mivalek.github.io/adata/marking/rubric_and_criteria_faq.html",
 
    flowchart_url = "https://raw.githubusercontent.com/mivalek/mivalek.github.io/master/adata/marking/AnD_report_flowchart.png",
  quick_comment_url = "https://mivalek.github.io/adata/marking/quick_comments.html",
  button_urls = "https://mivalek.github.io/adata/marking",
  height = "maximize"
)
}
\arguments{
\item{file}{Either path to .Rmd file to mark or, if working directory is set to the folder containing the file, ID number.}

\item{rubric_url}{\code{character}. A named list containing the rubric criteria (c1, c2, etc). Each criterion should be a list with \code{name}, \code{col}, and \code{text} elements containing the name, colour, and description of the criterion, respectively.}

\item{fdbck_boiler_text}{\code{character}. Optional path to boilerplate text to be used in the Feedback sections. See \code{teachR::mark()}}

\item{rubric_btn_url}{\code{cahracter}. URL to a webpage containing markng rubric/criteria/guidelines for markers' quick reference.}

\item{flowchart_url}{\code{cahracter}. URL to a webpage containing report flowchart students were supposed to follow. Also for for markers' quick reference. Can be either a .html file or image.}

\item{quick_comment_url}{\code{cahracter}. URL to a (Google?) spreadsheet containing quick comments for convenient marking.}
}
\description{
Wrapper for \code{teachR::mark()} producing a temporary preview of commented HTML file in the RStudio Viewer pane.
}
\details{
Function knits the file in the temp directory (\code{tempdir()} and opens it in RStudio's Viewer pane. Neither gubric grades nor marks are rendered. It requires all packages used in the Rmd report to be installed beforehand.
}
\examples{
preview(file = "marking/AnD/2019-20/48323.Rmd")
}
