% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{production}
\alias{production}
\title{Production Data Set}
\format{
A \code{\link[tibble]{tibble}} with 10 observations of 2 variables:
\describe{
  \item{x}{Something.}
  \item{y}{Something else.}
}
}
\source{
???
}
\usage{
production
}
\description{
A data set about production.
}
\keyword{datasets}
