% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_bank.R
\name{question.bank}
\alias{question.bank}
\title{Create a question bank from .Rmd exam item files.}
\usage{
question.bank(question_files, authors)
}
\arguments{
\item{question_files}{\code{character}. Paths to Rmd files with items. Best to have one file per author if there are multiple authors of exam items. See mcq_example.Rmd.Rmd in package folder for the correct markdown.}

\item{authors}{\code{character}. Requires one author per Rmd file.}
}
\value{
Function returns a \code{data.frame} with following columns:
    \code{q_num} Original item number in the Rmd file
    \code{week} Week of course/module the item pertains to
    \code{diff} Difficulty (Simple, Easy, Medium, Hard)
    \code{author} Author of the item
    \code{stem} Stem/lead-in
    \code{opts} Response options
    \code{correct} Correct answer
    \code{group} If item is a part of a group tht share a group stem, ID number identifying all items from the same group
    \code{group_stem} Stem shared by a group of items.
    
Week, and difficulty are extracted from the question tag, e.g. {02S} for week 2, simple. See mcq_example.Rmd in package folder for the correct markdown.
}
\description{
Produces a data frame that can be used as first argument to \code{\link{make.paper}()}.
}
\examples{
question.bank("exam/and_qu_MV.Rmd", "MV")
}
