% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raText.R
\name{raText}
\alias{raText}
\title{Grade open-ended answers}
\usage{
raText()
}
\value{
Three .csv files: one with students' points, one with updated solutions, one with updated criteria.
}
\description{
Shiny gadget to facilitate grading open-ended questions and essays.
}
\author{
Nicolas Mangin
}
