% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_parameters.R
\name{retrieve_parameters}
\alias{retrieve_parameters}
\title{Retrieve parameters}
\usage{
retrieve_parameters(ex_name, pkgname)
}
\arguments{
\item{ex_name}{Character. Name of the .Rmd file being processed.}

\item{pkgname}{Character. Name of the package containing the .Rmd file being processed.}
}
\value{
Parameters for the questions.
}
\description{
Get the question parameters when it is produced for a test.
}
\author{
Nicolas Mangin
}
