% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{schoice}
\alias{schoice}
\title{Single choice question}
\usage{
schoice(opts, sol)
}
\arguments{
\item{opts}{The options to choose from (character vector)}

\item{sol}{The correct solution (the correct value from \code{opts}, or the
position of the correct value)}
}
\description{
Use this function in a cloze inline chunk to describe a single choice
question.
}
