% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendexam.R
\name{sendexam}
\alias{sendexam}
\title{Shiny gadget to create and send individual reports to students.
Launch the gadget, upload the results exported from the gadget gradexam,
look at students' profiles, and send individual reports to the mailing list.}
\usage{
sendexam()
}
\description{
Shiny gadget to create and send individual reports to students.
Launch the gadget, upload the results exported from the gadget gradexam,
look at students' profiles, and send individual reports to the mailing list.
}
\seealso{
genexam()

gradexam()

checkexam()
}
