% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulanswers.R
\name{simulanswers}
\alias{simulanswers}
\title{Function creating a test of the exam generator based on the solutions recorded in an rds file.}
\usage{
simulanswers(exam_id = NULL, correction = 1/(5 - 1), nbr = 50,
  path = getwd())
}
\arguments{
\item{exam_id}{character. ID of the exam.}

\item{correction}{numeric. Penalty to apply for guesses.}

\item{nbr}{integer. Number of students to simulate.}

\item{path}{character. Path to the .rds file direcgtory.}
}
\value{
A list with two elements: answers from students and corresponding grades
}
\description{
Function creating a test of the exam generator based on the solutions recorded in an rds file.
}
