% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_to_pdf.R
\name{slides_to_pdf}
\alias{slides_to_pdf}
\title{Generate PDF lecture slides for a HUGO tachR website}
\usage{
slides_to_pdf(lecture_number, module = "AnD", port = 4321)
}
\arguments{
\item{lecture_number}{Number of the lecture to convert to PDF. May be a vector.}

\item{module}{\code{character}. Module name to use as a prefix in PDF file name. Defaults to \code{"AnD"}.}

\item{port}{Port number on which to locally serve the Hugo website hosting lecture slides.}
}
\description{
Requires a local install of \href{https://github.com/astefanutti/decktape}{decktape}. Function converts one or more knitted teachR::xaringan_slides files to a PDF version with slides next to a "notepad" sheet (\href{https://and.netlify.app/lectures/01/AnD_lecture01_slides.pdf}{example}).
}
\details{
Function will first check if \code{\link[blogdown:serve_site]{blogdown::serve_site()}} has been called. If it hasn't it will call it before running decktape on lectures hosted at \code{"http://localhost:[port]/lectures/[lecture_number]/slides/?handout=true"}.

Ouptu PDF gets saved to \code{"content/lectures/[lecture_number]/[module]_lecture[lecture_number]_slides.pdf"}.
}
\examples{
slides_to_pdf(1:3)

}
