% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidify.R
\name{slidify}
\alias{slidify}
\title{Convert .Rmd to reveal.js slides}
\usage{
slidify(file, course, meta_folder, incremental = FALSE, fig_width = 5,
  fig_height = 3.5, transition = "fade",
  background_transition = transition, plugins = c("notes", "search",
  "chalkboard"))
}
\arguments{
\item{file}{character. Path to .Rmd file to convert.}

\item{course}{character. Course the lecture is for: one of "dapR_1", "daprR_2", "dapR_3", "usmr", "msmr", "other".}

\item{meta_folder}{character. path to folder with slides.css, slides.js, and _header.html files for generating reveal.js slides. By default a subdirectory of directory where file= is located.}

\item{incremental}{logical. TRUE to render slide bullets incrementally on click. FALSE by default.}

\item{fig_width, fig_height}{numeric. numeric. Default width and height (in inches) for figures.}

\item{transition, background_transition}{character. Slide transition animation. "fade" by default. See reveal.js documentation for more options.}

\item{plugins}{character. which plugind to include. By default c("notes", "search", "chalkboard").}
}
\value{
Function does not return anything but outputs a .html file called [file]_slides.html and a corresponding folder with figures.
}
\description{
Lecture slides should be written in R Markdown with title, subtitle, and author in YAML header and saved on a local drive (due to Pandoc's issues with network drives). slidify() converts lecture .Rmd file into a reveal.js html presentation applying the theme for given course.
}
\examples{
# .Rmd and meta.folder must be on a local drive!
slidify("C:/Users/mvalasek/slides/dapR_1_demo.Rmd", "dapR_1", "C:/Users/mvalasek/meta")
}
