% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_compute_irt_parameters.R
\name{statistics_compute_irt_parameters}
\alias{statistics_compute_irt_parameters}
\title{Compite IRT parameters}
\usage{
statistics_compute_irt_parameters(predictions)
}
\arguments{
\item{predictions}{Tibbles.}
}
\value{
Tibble.
}
\description{
Compute difficulty, discrimination, and guess parameters
}
\author{
Nicolas Mangin
}
