% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_get_parameters.R
\name{statistics_get_parameters}
\alias{statistics_get_parameters}
\title{Compute various statistics}
\usage{
statistics_get_parameters(
  results,
  model_formula = "correct ~ success + proficiency",
  minobs = 10
)
}
\arguments{
\item{results}{Tibble. observation (e.g. student attempt in test), code (e.g. question) and whether the answer is correct.}

\item{model_formula}{Character Model used to estimate parameters (converted into a formula in yhe function)}

\item{minobs}{Integer.}
}
\value{
Models and parameters
}
\description{
Compute various statistics
}
\author{
Nicolas Mangin
}
