% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_update_ui.R
\name{statistics_update_ui}
\alias{statistics_update_ui}
\title{Update documents and data}
\usage{
statistics_update_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\description{
Module facilitating the update of underlying course documents, trees, data, and statistics.
}
\author{
Nicolas Mangin
}
