% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sussex_boilerplate.R
\name{sussex.boilerplate}
\alias{sussex.boilerplate}
\title{Generate boilerplate title page for Psychology@Sussex exams.}
\usage{
sussex.boilerplate(
  module,
  module_code,
  exam_diet,
  time_allowed,
  n_sections,
  marks_per_q,
  section_weights = NULL,
  sample = F
)
}
\arguments{
\item{module}{@param module_code \code{character}. Name and code of the module.}

\item{exam_diet}{\code{character}. Examinatiopn diet (e.g., sit - A2, resit - A3...). See example.}

\item{time_allowed}{\code{character}. Total time allowed for completion.}

\item{n_sections}{\code{numeric}. Number of sections in the paper.}

\item{marks_per_q}{\code{numeric}. Number of marks per item.}

\item{section_weights}{\code{character}. Line to include with information about section weighting.}

\item{sample}{\code{logical}. Should this be a sample paper title page? \code{FALSE} by default.}

\item{question_files}{\code{character}. Paths to Rmd files with items. Best to have one file per author if there are multiple authors of exam items. See example_items.Rmd in package folder for the correct markdown.}
}
\value{
Function returns a \code{data.frame} with following columns:
    \code{title_page} Text of the title page
    \code{yaml} YAML front matter
    
Week, and difficulty are extracted from the question tag, e.g. {02S} for week 2, simple. See example_items.Rmd in package folder for the correct markdown.
}
\description{
Produces a title page and yaml front matter text to be used in \code{\link{make.paper}()}.
}
\examples{
boiler <- sussex.boilerplate(module = "Analysing Data",
                             module_code = "C8891",
                             exam_diet = "BSc FIRST YEAR EXAMINATION May/June 2020 (A2)",
                             time_allowed = "2 hours",
                             n_sections = 1,
                             marks_per_q = 2,
                             section_weights = NULL,
                             sample = F)
}
