\name{teachR-package}
\alias{teachR-package}
\alias{teachR}
\docType{package}
\title{
\packageTitle{teachR}
}
\description{
\packageDescription{teachR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{teachR}
\packageIndices{teachR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{teachR}

Maintainer: \packageMaintainer{teachR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
