% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website.R
\name{teachr_website}
\alias{teachr_website}
\title{R Markdown site generator for teachr workshops}
\usage{
teachr_website(input, encoding = getOption("encoding"), ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory).}

\item{encoding}{Ignored. The encoding is always assumed to be UTF-8.}

\item{...}{Currently unused.}
}
\description{
R Markdown site generator for teachr workshops
}
