% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_calibrate.R
\name{test_calibrate}
\alias{test_calibrate}
\title{Analyze and weight criteria}
\usage{
test_calibrate()
}
\value{
Export updated weights and scores
}
\description{
Analyze item characteirtics to guide the weighting of criteria
}
\author{
Nicolas Mangin
}
