% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_generate.R
\name{test_generate}
\alias{test_generate}
\title{Generate Tests}
\usage{
test_generate()
}
\value{
Create an exam folder will all the the necessary files fo test administration and subsequent steps.
}
\description{
Shiny gadget to select test questions and generate both tests and solutions for several delivery formats.
}
\author{
Nicolas Mangin
}
