% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_func.R
\name{unlibrary}
\alias{unlibrary}
\title{Functions used for marking and within \code{\link{mark}()}}
\usage{
unlibrary(x = default_pkgs)
}
\description{
Functions used for marking and within \code{\link{mark}()}
}
