% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_people.R
\name{update.people}
\alias{update.people}
\title{Update People section of a course page}
\usage{
\method{update}{people}(surname = "all", course = "all", file,
  remove = F, inside = F)
}
\arguments{
\item{surname}{character. Either "all" (default) or surname of person to add to course.}

\item{course}{character. Course to modify: either "all" (default) or one of "dapR_1", "daprR_2", "dapR_3", "usmr", "msmr".}

\item{file}{character. Name of .csv file containing details (with path). Not required if remove = TRUE.}

\item{remove}{logical. TRUE removes people from page, FALSE adds them.}

\item{inside}{logical. Suppresses dialog windows. TRUE mainly for use within functions.}
}
\value{
Function does not return anything of use.
}
\description{
Function reqires an up-to-date website_people.csv file with all info. For more details about the file, see internal guide.
}
\details{
IMPORTANT: Set working directory to website folder!.
}
\examples{
# update all people on all courses
update.people(file = "M:/teaching/admin/website_people.csv")

# remove single person from all courses
update.people("valasek", remove = T)
}
