% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_answers.R
\name{update_answers}
\alias{update_answers}
\title{Update answers to tests}
\usage{
update_answers(course_paths)
}
\arguments{
\item{course_paths}{List. Paths to the different folders and databases on local disk.}
}
\description{
Function compiling all students' answers into a single database.
}
\author{
Nicolas Mangin
}
