% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_canvas.R
\name{update_canvas}
\alias{update_canvas}
\title{Update Canvas content}
\usage{
update_canvas()
}
\description{
Updates Canvas content by automatically searching through course materials in
the various repositories.
}
