% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_comments.R
\name{update_comments}
\alias{update_comments}
\title{Update comments}
\usage{
update_comments(course_paths)
}
\arguments{
\item{course_paths}{List. Paths to the different folders and databases on local disk.}
}
\description{
Function listing comments and gathering all information about them.
}
\author{
Nicolas Mangin
}
