% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update course tests}
\usage{
update_data(course_paths)
}
\arguments{
\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\description{
Function listing tests and gathering tests definition parameters in a single database which is saved on disk.
}
\author{
Nicolas Mangin
}
