% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_logs.R
\name{update_logs}
\alias{update_logs}
\title{Update logs}
\usage{
update_logs(course_paths)
}
\arguments{
\item{course_paths}{List. Paths to the different folders and databases on local disk.}
}
\description{
Function listing logs and gathering all information about them.
}
\author{
Nicolas Mangin
}
