% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_paths.R
\name{update_paths}
\alias{update_paths}
\title{Update student list}
\usage{
update_paths(course_paths)
}
\arguments{
\item{course_paths}{List. Paths to the different folders and databases on local disk.}
}
\description{
Function listing all students and gathering all information about them.
}
\author{
Nicolas Mangin
}
