% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ratings.R
\name{update_ratings}
\alias{update_ratings}
\title{Update ratings}
\usage{
update_ratings(course_paths)
}
\arguments{
\item{course_paths}{List. Paths to the different folders and databases on local disk.}
}
\description{
Function listing ratings and gathering all information about them.
}
\author{
Nicolas Mangin
}
