% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_statistics.R
\name{update_statistics}
\alias{update_statistics}
\title{Compte statistics for documents and items.}
\usage{
update_statistics(course_paths, minobs = 10)
}
\arguments{
\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{minobs}{Integer.}
}
\value{
Aggregates, models, and parameters saved on local folders.
}
\description{
Compte statistics for documents and items.
}
\author{
Nicolas Mangin
}
