% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tags.R
\name{update_tags}
\alias{update_tags}
\title{Update tags}
\usage{
update_tags(course_paths)
}
\arguments{
\item{course_paths}{List. Paths to the different folders and databases on local disk.}
}
\value{
Save the tag list as a tibble in the folder "2_documents" and returns this list
}
\description{
Function taking all the tags listed in documents, adding the new ones to the existing list of tags, and counting their occurrences.
}
\author{
Nicolas Mangin
}
