% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tests.R
\name{update_tests}
\alias{update_tests}
\title{Update test list}
\usage{
update_tests(course_paths)
}
\arguments{
\item{course_paths}{List. Paths to the different folders and databases on local disk.}
}
\description{
Function listing all tests and gathering all information about their content.
}
\author{
Nicolas Mangin
}
