% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_views.R
\name{update_views}
\alias{update_views}
\title{Update views}
\usage{
update_views(course_paths)
}
\arguments{
\item{course_paths}{List. Paths to the different folders and databases on local disk.}
}
\description{
Function listing views and gathering all information about them.
}
\author{
Nicolas Mangin
}
