% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_button.R
\name{week.button}
\alias{week.button}
\title{Add or remove a button on course content page}
\usage{
week.button(course, sem = NULL, label, link, remove = FALSE)
}
\arguments{
\item{course}{\code{character}. Course page to modify: one of \code{"dapR_1", "daprR_2", "dapR_3", "usmr"}, or \code{"msmr"}.}

\item{sem}{\code{numeric}. Semester to modify: either \code{1} or \code{2}. Only applicable for 2-semester courses, currently dapR 1 and 2.}

\item{label}{\code{character}. Button text. Keep it short, e.g., \code{"Lecture", "Lab", "Explainer"}...}

\item{link}{\code{character}. Single URL for the buttons. Defaults to /course/sem. Probably better to use \code{\link{link.button}()}.}

\item{remove}{\code{logical}. If \code{FALSE} (default), button gets added. If \code{TRUE} it gets removed}
}
\value{
Function does not return anything of use.
}
\description{
A button gets added/removed for each week of a given course/semester.
}
\details{
IMPORTANT: Set working directory to website folder!.
}
\examples{
# add Explainer button for each week
week.button(course = "dapR_1", sem = 2, label = "Explainer")

# remove Lecture button
week.button("msmr", label = "Lecture", rmeove = T)
}
