% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibRate.R
\name{calibRate}
\alias{calibRate}
\title{Shiny gadget to change and update the structure of a package.}
\usage{
calibRate()
}
\value{
Make necessary entries in structure databases.
}
\description{
Shiny gadget to change and update the structure of a package.
}
