% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_test_ui.R
\name{course_test_ui}
\alias{course_test_ui}
\title{Select and load a test}
\usage{
course_test_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A list co ntaining a test and a test folder path.
}
\description{
Module facilitating the selection of tests and tests to edit and work with.
}
\author{
Nicolas Mangin
}
