% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_tree_ui.R
\name{course_tree_ui}
\alias{course_tree_ui}
\title{Select and load a tree}
\usage{
course_tree_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
A list co ntaining a tree and a test folder path.
}
\description{
Module facilitating the selection of trees and tests to edit and work with.
}
\author{
Nicolas Mangin
}
