% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assignment_canvas.R
\name{create_assignment_canvas}
\alias{create_assignment_canvas}
\title{Create a new assignment on Canvas}
\usage{
create_assignment_canvas(name, number, week)
}
\arguments{
\item{name}{Assignment name as a character string. If left blank, assignments
are titled as 'Assignment (number)'}

\item{number}{Assignment number as an integer value.}

\item{week}{Week/module number for the assignment. The assignment is then
automatically placed in the correct location on the course website.}
}
\description{
Internal function. Creates a new assignment for the course on Canvas.
}
