% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_course.R
\name{create_course}
\alias{create_course}
\title{Create a new course}
\usage{
create_course(name, organization)
}
\arguments{
\item{name}{Course title as a character string. Example: 'DSCI 100 - Introduction to Data Science'}

\item{organization}{Name of the organization as a character string with no spaces.
This name must be available on GitHub, otherwise an error will be returned.
Example: 'dsci-100-2022'}
}
\value{

}
\description{
Creates a new GitHub organization for a course along with several repositories in the GitHub organization.
}
