% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_website.R
\name{initialize_website}
\alias{initialize_website}
\title{Initialize website for 'teachr' course}
\usage{
initialize_website(name)
}
\arguments{
\item{name}{See create_course()}
}
\description{
Internal function. Creates the website. Only run this function once.
}
\keyword{internal}
