% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_canvas.R
\name{link_canvas}
\alias{link_canvas}
\title{Link to a Canvas course}
\usage{
link_canvas(token_path, course_id, API_URL)
}
\arguments{
\item{token_path}{Path to where the Canvas API token number is stored. This
should be stored in a file of type '.txt'. The Canvas token can be obtained in
Settings > Approved Integrations > New Access Token.}

\item{course_id}{Canvas course number as an integer.}

\item{API_URL}{URL for Canvas API}
}
\description{
Establish a connection with the Canvas Instructure LMS API. Doing so allows
for automatic updates to the Canvas course. This only needs to be performed once.
}
