% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_apps.R
\name{list_apps}
\alias{list_apps}
\title{List \code{teachR} Apps}
\usage{
list_apps(categories = NULL, keywords = NULL)
}
\arguments{
\item{categories}{A character string or a vector of character strings 
corresponding to existing app categories (case insensitive).}

\item{keywords}{A character string or a vector of character strings 
corresponding to existing app keywords (case insensitive).}
}
\value{
This function opens an html file.
}
\description{
This function generates a webpage listing all the apps available 
 through the \code{teachR} library. The list can be filtered by category 
 (type \code{list_categories()} for a list of existing categories) and 
 keywords (type \code{list_keywords()} for a list of existing keywords).
}
\details{
If called from RStudio, the list will open in the internal RStudio 
 viewer. If called from a terminal, it will open in your default internet 
 browser.
}
\examples{
\dontrun{
 list_apps(categories = c("biology", "social science"))
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
