% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_categories.R
\name{list_categories}
\alias{list_categories}
\title{List Categories of All \code{teachR} Apps}
\usage{
list_categories()
}
\value{
A vector of character strings, each of them corresponding to a different
 category.
}
\description{
This function generates a listing  of all the categories of apps 
 available through the \code{teachR} library.
}
\examples{
\dontrun{
 list_categories()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
