% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_keywords.R
\name{list_keywords}
\alias{list_keywords}
\title{List Keywords of All \code{teachR} Apps}
\usage{
list_keywords()
}
\value{
A vector of character strings, each of them corresponding to a different
 keyword.
}
\description{
This function generates a listing of all the keywords of apps 
 available through the \code{teachR} library.
}
\examples{
\dontrun{
 list_keywords()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
