% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sheet.R
\name{make.sheet}
\alias{make.sheet}
\title{Convert .Rmd to course-formatted HTML document or R Notebook}
\usage{
make.sheet(
  file,
  course,
  solution = F,
  handout = FALSE,
  tasks_to_headings = FALSE,
  notes = TRUE,
  ntb = FALSE,
  color = NULL,
  toc = T,
  toc_depth = 2,
  toc_float = T,
  fig_width = 5,
  fig_height = 3.5,
  highlight = "tango",
  colour = color,
  ...
)
}
\arguments{
\item{file}{\code{character}. Path to .Rmd file to convert.}

\item{course}{\code{character}. Course the sheet is for: one of \code{"dapR_1", "daprR_2", "dapR_3", "usmr", "msmr", "other"}.}

\item{solution}{\code{logical}. Should solutions to taks be rendered? \code{FALSE} by default.}

\item{handout}{\code{logical}. \code{TRUE} adds "Click for slides" link under author. To be used only for slide handout HTML files. \code{FALSE} by default.}

\item{tasks_to_headings}{\code{logical}. Should individual tasks and subtaks be rendered as level 2 and 3 (respectively) headings? \code{FALSE} by default.}

\item{notes}{\code{logical}. If \code{handout=TRUE}, then if \code{TRUE} a writable box for notes will appear at the bottom of each slide.}

\item{ntb}{\code{logical}. \code{TRUE} to render document as R Notebook. \code{FALSE} by default.}

\item{color}{\code{character}. Either a single valid colour (hex code or any of the values in \code{colours()}) or any valid value of the \code{course=} argument. If provided, it will be used to set colour scheme instead of \code{course=}.}

\item{toc}{\code{logical}. Should table of content be included}

\item{toc_depth}{\code{logical}. Depth of headers to include in table of contents.}

\item{toc_float}{\code{TRUE} to float the table of contents to the left of the main document content. Rather than TRUE you may also pass a list of options that control the behavior of the floating table of contents. For more details, see \code{\link[rmarkdown]{html_document}}.}

\item{fig_width, fig_depth}{\code{numeric}. Default width and height (in inches) for figures.}

\item{highlight}{\code{character}. Syntax highlighting style. See \code{\link[rmarkdown]{html_document}}.}

\item{...}{Other arguments to be passed to \code{rmarkdown:html_document} or \code{rmarkdown:html_notebook}.}
}
\value{
\code{TRUE} if output .html file was successfully created.
}
\description{
Lab sheets and lecture materials should be written in R Markdown with title, subtitle, and author in YAML header. \code{make.sheets()} converts a .Rmd file into a HTML/R Notebook document applying the theme for given course.
}
\details{
Function requires a .css and .js files for correct formatting of lab sheets/handouts. These files sit on the stats website in the [root]/sheet_files folder and the path is hard-coded into the function. Look for css and js objects in function body.
}
\examples{
make.sheet("C:/Users/mvalasek/slides/dapR_1_handout_demo.Rmd", "dapR_1")
}
\seealso{
\code{\link{handout}()}
}
