% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark}
\alias{mark}
\title{Mark coursework assignment Rmd}
\usage{
mark(
  file,
  file_name = file,
  study = NULL,
  mark = NULL,
  rubric_grades = NULL,
  rubric = NULL,
  include_rubric_desc = F,
  feedback = F,
  count_words = !feedback,
  limit = 2000,
  include_results = F,
  results_obj = NULL,
  remove_results = F,
  color = "#b38ed2"
)
}
\arguments{
\item{file}{\code{character}. Either path to .Rmd file to mark or URL from which to read the file.}

\item{file_name}{\code{character}. If \code{file=} is a URL, name of file to write to.}

\item{study}{\code{character}. Which study to get correct results for? Currently either \code{"red"} or \code{"green"}.}

\item{mark}{\code{numeric}. Total mark/grade for given assignment.}

\item{rubric}{A named list containing the rubric criteria (c1, c2, etc). Each criterion should be a list with \code{name}, \code{col}, and \code{text} elements containing the name, colour, and description of the criterion, respectively.}

\item{include_rubric_desc}{\code{logical}. If \code{TRUE} and if \code{rubric_grades} is a factor, grade descriptors for each marking criterion will be included in sidebar boxes.}

\item{feedback}{\code{logical}. Should feedback box with marker's comments appear at the bottom of document? \code{FALSE} by default.}

\item{count_words}{\code{logical}. \code{TRUE} counts body text words and if word count exceeds \code{limit}, inserts a warning line in HTML. \code{!feedback} by default.}

\item{limit}{\code{numeric}. Word count limit}

\item{include_results}{\code{logical}. If \code{TRUE}, object provided to \code{results_obj=} will get inserted into a formatted box in knitted HTML. \code{FALSE} by default.}

\item{results_obj}{See \code{include_results}.}

\item{remove_results}{\code{logical}. If \code{TRUE}, correct results will not be displayed in final marked HTML document. \code{FALSE} by default.}

\item{color}{\code{character}. Single valid colour string (hex code or any of the values in \code{colours()}).}

\item{rubric_grades.}{Mark for each rubric criterion. Can be a factor with levels sorted from lowest mark to highest mark. See \code{rubric} and \code{include_rubric_desc} for more info.}
}
\value{
\code{TRUE} if output .Rmd and .html file were successfully created.
}
\description{
For Rmd submissions of reports and other coursework. Function checks word count and inserts space for feedback.
}
\details{
Function run with \code{feedback = FALSE} will overwrite the original Rmd file and knit it into HTML. This should be done before marking in order to automate word count. The overwritten Rmd should then be used for comments and feedback. Once done, \code{mark(feedback = TRUE)} should be run on the edited Rmd. This will output a ..._marked.html file that can be returned to students.

In-text comments can be inserted into .Rmd file on a lew line surrounded by the <fb></fb> HTML tag.
}
\examples{
# first run
mark("C:/work/201000.Rmd")
# then
mark("C:/work/201000.Rmd", T)
}
