% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_quest.R
\name{param_quest}
\alias{param_quest}
\title{Function to retrieve or generate parameters used when questions are generated.}
\usage{
param_quest(wdir = "", question_id = "", alttype = NULL, altlevel = NULL)
}
\arguments{
\item{wdir}{Character. Working directory.}

\item{question_id}{Character. ID of the question.}

\item{alttype}{Character. force the type of question: "mcq" or "open"}

\item{altlevel}{Character. force the type of level for the question: "2 Understand" or "3 Apply"}
}
\value{
Parameters for the questions.
}
\description{
Function to retrieve or generate parameters used when questions are generated.
}
