% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_parameters.R
\name{retrieve_parameters}
\alias{retrieve_parameters}
\title{Retrieve parameters}
\usage{
retrieve_parameters(wdir = "", questionid = "")
}
\arguments{
\item{wdir}{Character. Path to working directory.}

\item{questionid}{Character. ID of the question.}
}
\value{
Parameters for the questions.
}
\description{
Get the question parameters when it is prodiced for a test.
}
\author{
Nicolas Mangin
}
