% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run teaching apps provided by \code{\link[<teachR>:<teachR>-package]{teachR}}}
\usage{
run_app(app)
}
\arguments{
\item{app}{The name of the app to run.}
}
\value{
This function does not return anything.
}
\description{
This function calls a Shiny app provided by the \code{\link[<teachR>:<teachR>-package]{teachR}}
library. The list of available Shiny apps can be obtained by typing
\code{list_apps()} in your R terminal.
}
\details{
If called from RStudio, the app will open in the internal RStudio
internet browser. If called from a terminal, it will open in your default
internet browser. All apps should work without any problem with the internal
RStudio internet browser, as well as with recent versions of most internet
browsers. It is likely to break with older versions.
}
\examples{
run_app("aggregation_segregation")
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

