% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidify.R
\name{slidify}
\alias{slidify}
\title{Convert .Rmd to reveal.js slides}
\usage{
slidify(file, course, meta_folder)
}
\arguments{
\item{file}{character. Path to .Rmd file to convert.}

\item{course}{character. Course the lecture is for: one of "dapR_1", "daprR_2", "dapR_3", "usmr", "msmr".}

\item{meta.folder}{character. path to folder with .css, .js, and _header.html files for generating reveal.js slides. By default a subdirectory of directory where file= is located.}
}
\value{
Function does not return anything but outputs a .html file called [file]_slides.html and a corresponding folder with figures.
}
\description{
Lecture slides should be written in R Markdown with output set as html_document and saved on a local drive (due to Pandoc's issues with network drives). slidify() converts lecture .Rmd file into a reveal.js html presentation applying the theme for given course.
}
\examples{
# .Rmd and meta.folde must be on a local drive!
slidify("C:/Users/mvalasek/slides/dapR_1_demo.Rmd", "dapR_1", "C:/Users/mvalasek/meta")
}
