% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidify.R
\name{slidify}
\alias{slidify}
\title{Convert .Rmd to reveal.js slides}
\usage{
slidify(file, course, header_text = "default", incremental = FALSE,
  offline = FALSE, offline_css, fig_width = 5, fig_height = 3.5,
  transition = "fade", background_transition = transition,
  plugins = c("notes", "search", "chalkboard"))
}
\arguments{
\item{file}{character. Path to .Rmd file to convert.}

\item{course}{character. Course the lecture is for: one of "dapR_1", "daprR_2", "dapR_3", "usmr", "msmr", "other".}

\item{header_text}{character. Text to be displayed in the top right corner of slides. Course-specific text by default.}

\item{incremental}{logical. TRUE to render slide bullets incrementally on click. FALSE by default.}

\item{offline}{logical. TRUE generates reaveal.js presentation that only works locally. FALSE by default. If TRUE, offline_css can be specified.}

\item{offline_css}{character. Path to .css file for offline presentations. If not specified, slides will be linked to default css hosted online and will require Internet connection even if hosted offline. Leave unspecified if offline = FALSE.}

\item{fig_width, fig_height}{numeric. numeric. Default width and height (in inches) for figures.}

\item{transition, background_transition}{character. Slide transition animation. "fade" by default. See reveal.js documentation for more options.}

\item{plugins}{character. which plugind to include. By default c("notes", "search", "chalkboard").}
}
\value{
Function does not return anything but outputs a .html file called [file]_slides.html and a corresponding folder with figures.
}
\description{
Lecture slides should be written in R Markdown with title, subtitle, and author in YAML header. slidify() converts lecture .Rmd file into a reveal.js html presentation applying the theme for given course.
}
\details{
Function requires a .css and .js files for correct formatting of lab sheets/handouts. These files sit on the stats website in the [root]/slides_files folder and the path is hard-coded into the function. Look for css and js objects in function body.
}
\examples{
slidify("C:/Users/mvalasek/slides/dapR_1_handout_demo.Rmd", "dapR_1")
}
