% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidify.R
\name{slidify}
\alias{slidify}
\title{Convert .Rmd to reveal.js slides}
\usage{
slidify(file, course, header_text = "default", incremental = FALSE,
  offline = FALSE, offline_css, fig_width = 5, fig_height = 3.5,
  transition = "fade", background_transition = transition,
  plugins = c("notes", "search", "chalkboard"))
}
\arguments{
\item{file}{\code{character}. Path to .Rmd file to convert.}

\item{course}{\code{character}. Course the lecture is for: one of \code{"dapR_1", "daprR_2", "dapR_3", "usmr", "msmr", "other"}.}

\item{header_text}{\code{character}. Text to be displayed in the top right corner of slides. Course-specific text by default.}

\item{incremental}{\code{logical}. \code{TRUE} to render slide bullets incrementally on click. \code{FALSE} by default.}

\item{offline}{\code{logical}. \code{TRUE} generates reaveal.js presentation that only works locally. \code{FALSE} by default. If \code{TRUE}, \code{offline_css=} can be specified.}

\item{offline_css}{\code{character}. Path to .css file for offline presentations. If not specified, slides will be linked to default CSS hosted online and will require Internet connection even if hosted offline. Leave unspecified if \code{offline = FALSE}.}

\item{link_to_ho} {\code{logical}. \code{TRUE} (default) to display "click for handout" on title slide.}

\item{fig_width, fig_height}{\code{numeric}. Default width and height (in inches) for figures.}

\item{transition, background_transition}{\code{character}. Slide transition animation. \code{"fade"} by default. See reveal.js documentation (https://github.com/hakimel/reveal.js/blob/master/README.md) for more options.}

\item{plugins}{\code{character}. which plugind to include. By default \code{c("notes", "search", "chalkboard")}.}
}
\value{
Function does not return anything but outputs a .html file called [file]_slides.html and a corresponding folder with figures.
}
\description{
Lecture slides should be written in R Markdown with title, subtitle, and author in YAML header. \code{slidify()} converts lecture .Rmd file into a reveal.js html presentation applying the theme for given course.
}
\details{
Function requires a .css and .js files for correct formatting of lab sheets/handouts. These files sit on the stats website in the [root]/slides_files folder and the path is hard-coded into the function. Look for css and js objects in function body.
}
\examples{
slidify("C:/Users/mvalasek/slides/dapR_1_handout_demo.Rmd", "dapR_1")
}
