% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_assign_colors.R
\name{statistics_assign_colors}
\alias{statistics_assign_colors}
\title{Assign colors.}
\usage{
statistics_assign_colors(parameters, type = "questions")
}
\arguments{
\item{parameters}{Tibble. Page ratings.}

\item{type}{Character. Whether the parameterds are for "ratings", "videos", or "questions".}
}
\value{
Parameters with variable indicating colors for value boxes.
}
\description{
Function assigning colors based on documents' statistics.
}
\author{
Nicolas Mangin
}
