% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachr-package.R
\docType{package}
\name{teachr-package}
\alias{teachr}
\alias{teachr-package}
\title{teachr: LLM Integration to Teach You R}
\description{
Get feedback on your R code and exercises, autograde exercises, create exercises.
}
\keyword{internal}
