% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_course_details.R
\name{update.course.details}
\alias{update.course.details}
\title{Update Course details section of a course page}
\usage{
\method{update}{course.details}(details.file = "default",
  people.file = "default")
}
\arguments{
\item{details.file}{\code{character}. Name of .csv file containing course details. By default \code{"//chss.datastore.ed.ac.uk/chss/ppls/shared/courses/DAP/website_courses.csv"}. See details for MacOS and Linux users.}

\item{people.file}{\code{character}. Name of .csv file containing people details. By default, replaces \code{"_courses.csv"} in string given to \code{details.file=} with \code{"_people.csv"}.}
}
\value{
Function does not return anything of use.
}
\description{
Function reqires an up-to-date website_courses.csv and website_people.csv files (both can be renamed) with all info within. For more details about the files, see internal guide.
}
\details{
IMPORTANT: Set working directory to website folder!.

The function will by default try to access //chss.datastore.ed.ac.uk/chss... This only works on Windows. If it cannot access the path, it will try /Volumes/chss... instead. This should work on MacOS. For Linux users, you're on your own but I bet you're used to that by now.
}
\examples{
# update course details for all courses
update.course.details()
}
