% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_course_details.R
\name{update.course.details}
\alias{update.course.details}
\title{Update Course details section of a course page}
\usage{
\method{update}{course.details}(details.file,
  people.file = sub("_courses.csv", "_people.csv", details.file))
}
\arguments{
\item{details.file}{character. Name of .csv file containing course details (with path).}

\item{people.file}{character. Name of .csv file containing people details (with path). By default, replaces "_courses.csv" in string given to details.file= with "_people.csv".}
}
\value{
Function does not return anything of use.
}
\description{
Function reqires an up-to-date website_courses.csv and website_people.csv files (both can be renamed) with all info within. For more details about the files, see internal guide.
}
\details{
IMPORTANT: Set working directory to website folder!.
}
\examples{
# update course details for all courses
update.course.details(details.file = "M:/teaching/admin/website_courses.csv")
}
