% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_people.R
\name{update.people}
\alias{update.people}
\title{Update People section of a course page}
\usage{
\method{update}{people}(surname = "all", course = "all",
  file = "default", remove = F, inside = F)
}
\arguments{
\item{surname}{\code{character}. Either \code{"all"} (default) or surname of person to add to course.}

\item{course}{\code{character}. Course to modify: either \code{"all"} (default) or one of \code{"dapR_1", "daprR_2", "dapR_3", "usmr", "msmr"}.}

\item{file}{\code{character}. Name of .csv file containing details. By default \code{"//chss.datastore.ed.ac.uk/chss/ppls/shared/courses/DAP/website_people.csv"}. Not required if \code{remove = TRUE}. See details for MacOS and Linux users.}

\item{remove}{\code{logical}. \code{TRUE} removes people from page, \code{FALSE} adds them.}

\item{inside}{\code{logical}. Suppresses dialog windows. \code{TRUE} mainly for use within functions.}
}
\value{
Function does not return anything of use.
}
\description{
Function reqires an up-to-date website_people.csv file with all info. For more details about the file, see internal guide.
}
\details{
IMPORTANT: Set working directory to website folder!.

The function will by default try to access //chss.datastore.ed.ac.uk/chss... This only works on Windows. If it cannot access the path, it will try /Volumes/chss... instead. This should work on MacOS. For Linux users, you're on your own but I bet you're used to that by now.
}
\examples{
# update all people on all courses
update.people()

# remove single person from all courses
update.people("valasek", remove = T)
}
