% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_documents.R
\name{update_documents}
\alias{update_documents}
\title{Update document list}
\usage{
update_documents(course_paths)
}
\arguments{
\item{course_paths}{List. Paths to the different folders and databases on local disk.}
}
\value{
Save the document list as a tibble in the folder "2_documents" and returns this list
}
\description{
Function listing all the documents, checking whether they were changed, and updating information about changed documents.
}
\author{
Nicolas Mangin
}
