% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_statistics.R
\name{update_statistics}
\alias{update_statistics}
\title{Compute statistics.}
\usage{
update_statistics(course_paths, minobs = 10)
}
\arguments{
\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{minobs}{Integer.}
}
\description{
Function computing various statistics for documents and items and saving them in disk.
}
\author{
Nicolas Mangin
}
