% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_trees.R
\name{update_trees}
\alias{update_trees}
\title{Update course trees}
\usage{
update_trees(course_paths)
}
\arguments{
\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\description{
Function adding missing documents as unclassified, removing non-existing documents in all trees, and saving changes on disk.
}
\author{
Nicolas Mangin
}
