% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_button.R
\name{week.button}
\alias{week.button}
\title{Add or remove a button on course content page}
\usage{
week.button(course, sem = NULL, label, link, remove = FALSE)
}
\arguments{
\item{course}{character. Course page to modify: one of "dapR_1", "daprR_2", "dapR_3", "usmr", or "msmr".}

\item{sem}{numeric. Semester to modify: either 1 or 2. Only applicable for 2-semester courses, currently dapR 1 and 2.}

\item{label}{character. Button text. Keep it short, e.g., "Lecture", "Lab", "Explainer"...}

\item{link}{character. Single URL for the buttons. Defaults to /course/sem. Probably better to use link.button().}

\item{remove}{logical. If FALSE (default), button gets added. If TRUE it gets removed}
}
\value{
Function does not return anything of use.
}
\description{
A button gets added/removed for each week of a given course/semester.
}
\details{
IMPORTANT: Set working directory to website folder!.
}
\examples{
# add Explainer button for each week
week.button(course = "dapR_1", sem = 2, label = "Explainer")

# remove Lecture button
week.button("msmr", label = "Lecture", rmeove = T)
}
